Suspend immigration - EO - Protecting the Nation.. (#100)
Sec. 3.  Suspension of Issuance of Visas and Other Immigration Benefits to Nationals of Countries of Particular Concern.  (a)  The Secretary of Homeland Security, in consultation with the Secretary of State and the Director of National Intelligence, shall immediately conduct a review to determine the information needed from any country to adjudicate any visa, admission, or other benefit under the INA (adjudications) in order to determine that the individual seeking the benefit is who the individual claims to be and is not a security or public-safety threat.