Presidential Executive Order on Promoting Energy Independence and Economic Growth
Sec. 3.  Rescission of Certain Energy and Climate-Related Presidential
and Regulatory Actions.  (a)  The following Presidential actions are
hereby revoked:

(i)    Executive Order 13653 of November 1, 2013 (Preparing the United
States for the Impacts of Climate Change);

(ii)   The Presidential Memorandum of June 25, 2013 (Power Sector
Carbon Pollution Standards);

(iii)  The Presidential Memorandum of November 3, 2015 (Mitigating
Impacts on Natural Resources from Development and Encouraging Related
Private Investment); and

(iv)   The Presidential Memorandum of September 21, 2016 (Climate
Change and National Security).

(b)  The following reports shall be rescinded:

(i)   The Report of the Executive Office of the President of June 2013
(The President's Climate Action Plan); and

(ii)  The Report of the Executive Office of the President of March 2014
(Climate Action Plan Strategy to Reduce Methane Emissions).

(c)  The Council on Environmental Quality shall rescind its final
guidance entitled "Final Guidance for Federal Departments and Agencies
on Consideration of Greenhouse Gas Emissions and the Effects of Climate
Change in National Environmental Policy Act Reviews," which is referred
to in "Notice of Availability," 81 Fed. Reg. 51866 (August 5, 2016).